feat(requests): unified RequestsFeed + RequestCard mirroring specialists UX #1659
- New components/requests/RequestsFeed.tsx: mode='catalog' (public bourse
  with CityFnsCascade filter + pagination) and mode='mine' (own requests
  with active/closed tab switcher), matching SpecialistFeed architecture
- app/(tabs)/requests.tsx → thin wrapper over RequestsFeed (was 320 LOC)
- app/requests/index.tsx → thin wrapper over RequestsFeed (was 331 LOC)
- app/(tabs)/my-requests.tsx → thin wrapper; removes inline duplicate RequestCard
- All three screens now share PageTitle, filter bar, card layout, empty/error states
- tsc --noEmit: 0 errors frontend + backend
